Fix export of CLI functions using NodeJS specific methods
