+++ clients/typescript/src/migrators/index.ts
@@ -2,7 +2,8 @@ import { Statement } from '../util'
 
 export { BundleMigrator } from './bundle'
 export { MockMigrator } from './mock'
-export { loadMigrations } from './builder'
+export { parseMetadata, makeMigration } from './builder'
+export type { MetaData } from './builder'
 
 export interface Migration {
   statements: string[]
