+++ clients/typescript/test/migrators/builder.test.ts
@@ -1,9 +1,9 @@
 import test from 'ava'
 import {
-  loadMigrations,
   makeMigration,
   parseMetadata,
 } from '../../src/migrators/builder'
+import { loadMigrations } from '../../src/cli'
 import {
   SatOpMigrate,
   SatOpMigrate_Table,
